Refactor infinite loop when there are no datasets
`most_recent_datasets` was trying to fill its quota without running an
overlarge query by getting the latest datasets a few at a time, filtering
out the private ones, and looping until the desired number were found.
But if there are no datasets, as on a fresh provision, that never happens
and the loop never terminates.
This changes it to use the `package_search` action, which supports
filtering and limiting directly.