Fix broken context references on dataset group/topic view
This was a case where the master branch is different from CKAN 2.8.1 and
I mistakenly updated our overrides based on the master version.

I did some searching in the code to try to see if there are other cases
of this type of error hiding out, and I think there aren't.

Resolves https://github.com/azavea/opendataphilly-ckan/issues/88