Show 'Metadata last updated' but no 'Last updated' on datasets
Since many or most of the resource entities on ODP datasets point to
external resources that could change at any time, the "last_modified"
field on resources doesn't reflect reality, and will tend to make things
look out-of-date when they actually aren't. So this removes "Last updated"
on the Resource view and changes "Last Updated" on the Dataset view to
say "Metadata last updated", to avoid confusion.